Adapt code to run for huge datasets

+++ src/dynamic/parser.h
@@ -4,7 +4,7 @@
 #include <string>
 
 struct cmd_args {
-    int batch_size = 0;
+    int64_t batch_size = 0;
     bool directed = false;
     bool weighted = false;
     int64_t num_nodes = 0;
